Move TPE module to new file
